Updated README.md
The "Routing" project is a simple React web application that showcases the implementation of client-side routing using React Router. It consists of various components like Header, Home, Footer, Contact, and Services, each serving a specific purpose on the web page.

The web application features a navigation bar that allows users to navigate between different sections of the site, such as the home page, contact page, and services page. The landing page presents a captivating main header with the title "SOULTECH" and a brief description of the company.

The contact page provides a user-friendly contact form where visitors can submit their name, email, and queries. This enables easy communication between the users and the company.

The services page showcases a carousel of images using the react-responsive-carousel library. It displays images representing various services offered by the company, providing an attractive visual presentation.

The project emphasizes simplicity and serves as an excellent starting point for beginners to explore React, React Router, and basic web development concepts. Developers can contribute to the project, improve its functionality, or use it as a foundation for building more complex web applications.